Fix a crititcal bug in the decision process
If the postgresql on the first node returned by the query to find
candidates in do_failover is down then the initialization of the
bestCandidate is done with non assigned variables.

Fix the situation by moving the initialization in the loop above.
And loop until we have a find_best. Added a log message if no candidate
is found